% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attacher-package.R
\docType{package}
\name{attacher}
\alias{attacher}
\alias{attacher-package}
\title{Attach graphs and figures}
\description{
Functions to facilitate attaching graphs and plots to pdf document created
with rnw files + knitr using the `attachfile`
}

