% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caption-functions.r
\name{fig_cap}
\alias{fig_cap}
\title{Caption generator}
\usage{
fig_cap(cap, attach = NULL, ext = NULL, warn = TRUE)
}
\arguments{
\item{cap}{Caption}

\item{attach}{attach graph? if \code{NULL} then this parameter will
be taken from \code{opts_proh$get('attach_graph')}}

\item{ext}{what is the file extension? if \code{NULL} then this parameter will
be taken from \code{opts_proh$get('graph_dev')}}

\item{warn}{warn if extension does not seem available?}
}
\description{
Use this with chunk option \code{fig.cap} to attach a copy
of the produced graphic in the resulting pdf.
}
\note{
Beware if you want different extension in the attachment than
  appears in your pdf. LaTeX has a default order in which it chooses
  formats if a file name is not unique (knitr does not specify extension).
  E.g. you might want pdf:s in your pdf but .png in you attachment. Then
  you can specify something like
  \\DeclareGraphicsExtensions\{.pdf,.eps,.png,.jpg,jpeg\} in the preamble
  of you LaTeX document.
}
