% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hook-functions.r
\name{hook_tab.cap}
\alias{hook_tab.cap}
\title{Chunk Tag \code{tab.cap} Extension for package:knitr}
\usage{
hook_tab.cap()
}
\description{
\pkg{knitr} hook functions are called when the corresponding
  chunk options are not \code{NULL} to do additional jobs beside the R code
  in chunks. \pkg{enumeratermd} provides the hook \code{tab.cap} which adds a
  Pandoc table caption after the chunk. It is meant to be analogous to chunk
  option \code{fig.cap}.
}
\details{
the function hook_tab.cap is set as a hook in \pkg{knitr} when
  \pkg{attacher} is attached (and removed when \pkg{attacher} is
  detached).This hook creates an \\begin{table} environment before
  the chunk output and a \\end{table} after the chunk output (+ label, etc.)
}
\references{
\url{http://yihui.name/knitr/hooks#chunk_hooks}
}

