% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-set-options.r
\docType{data}
\name{opts_attacher}
\alias{opts_attacher}
\title{attacher options}
\format{An object of class \code{list} of length 3.}
\usage{
opts_attacher
}
\description{
This list tries to mimic the behaviour of opts_chunk from knitr.
Currently these values are maintained with the functions in (the list)
\code{opts_attacher}:
\itemize{
\item attach_graph - use \code{fig_cap} in \code{fig.caption}
(in a \code{knitr} chunk), i.e. \code{fig.caption = fig_cap("My Caption")},
this forces \code{fig_cap} to include an attach-statement
pointing to the relevant figure in 'figure/' .
\item attach_table - if TRUE attached tables in caption created
    with \code{tab_cap} (default FALSE). You can use the added hook
    'tab.cap' (analogous to knitr 'fig.cap'), i.e. \code{tab.cap =
    tab_cap("My caption", ...)}.
\item table_path a subdirectory in which to store table files (default "attacher_tables")
\item graph_dev - graphical extension (default "pdf")
\item table_ext the extension for the table files (default "csv")
\item table_fnc - this is the function to write the \code{object} argument in \code{tab_cap}
to file (default \code{utils::write.csv})
}
}
\keyword{datasets}

