% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hook-functions.r
\name{hook_tab.cap}
\alias{hook_tab.cap}
\title{Chunk Tag \code{tab.cap} Extension for package:knitr}
\usage{
hook_tab.cap()
}
\description{
\pkg{knitr} hook functions are called when the corresponding
  chunk options are not \code{NULL} to do additional jobs beside the R code
  in chunks. \pkg{eattacher} provides the hook \code{tab.cap} and
  \code{tab.scap} which are meant to be analogous to \code{fig.cap} and
  \code{fig.scap}, respectively.
}
\details{
the function hook_tab.cap is set as a hook in \pkg{knitr} when
  \pkg{attacher} is attached (and removed when \pkg{attacher} is
  detached).This hook creates an \\begin{table} environment before the chunk
  output, as well as a caption,  and a \\end{table} after the chunk output (+
  label, etc.) The hook_tab.scap shortens the caption for the 'lot'.
}
\references{
\url{http://yihui.name/knitr/hooks#chunk_hooks}
}
