% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-set-options.r
\docType{data}
\name{opts_attacher}
\alias{opts_attacher}
\title{attacher options}
\format{\preformatted{List of 3
 $ get    :function (name)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 13 17 25 1 17 1 13 25
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000159e06f8> 
 $ set    :function (...)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 31 17 40 1 17 1 31 40
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000159e06f8> 
 $ restore:function ()  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 45 21 58 1 21 1 45 58
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x00000000159e06f8> 
}}
\usage{
opts_attacher
}
\description{
This list tries to mimic the behaviour of opts_chunk from knitr.
Currently these values are maintained with the functions in (the list)
\code{opts_attacher}:
\itemize{
\item attach_graph - use \code{figh} in \code{fig.caption}
(in a \code{knitr} chunk), i.e. \code{fig.caption = figh("My Caption")},
this forces \code{figh} to include an attach-statement
pointing to the relevant figure in 'figure/' .
\item attach_table - this .... needs documentation
\item table_path a subdirectory in which to store table files
\item graph_dev - graphical extension
\item table_ext the extension for the table files
\item table_fnc - this is the function to write the \code{tableh}-table
to file
}
}
\keyword{datasets}

